feat(e2ei): filter out root CA when registering intermediates in case the provider repeats it
